requirements.txt
google-generativeai
langchain
langchain-google-genai
langchain-community
pymongo
langchain-mongodb
python-dotenv
PyPDF
langchain-huggingface
sentence-transformers
